CoffeeShop-Web-application/CoffeeMenu/backend/requirements.txt
astroid==2.2.5
Click==7.0
ecdsa==0.13.2
Flask==1.0.2
Flask-SQLAlchemy==2.5.0
future==0.17.1
isort==4.3.18
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.0
MarkupSafe==1.1.1
mccabe==0.6.1
pycryptodome==3.3.1
pylint==2.3.1
python-jose-cryptodome==1.3.2
six==1.12.0
typed-ast==1.4.2
Werkzeug==0.15.6
wrapt==1.11.1
Flask-Cors==3.0.8

Manager: eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCIsImtpZCI6IlJ2d001NV9jdzEwdTZ5Sk9aNlMwNCJ9.eyJpc3MiOiJodHRwczovL2NvZmZlZS1kZXYudXMuYXV0aDAuY29tLyIsInN1YiI6ImF1dGgwfDYyYmMyZmVkOGFmZTdiMDdjNjlkMzBmZCIsImF1ZCI6ImNvZmZlZXNob3AiLCJpYXQiOjE2NTY1MDA5MTAsImV4cCI6MTY1NjUwODExMCwiYXpwIjoiVzZ1V2hIYkt5RDdmSzE5YzRYS0JHTHROTDdDaHFha2kiLCJzY29wZSI6IiIsInBlcm1pc3Npb25zIjpbImRlbGV0ZTpkcmlua3MiLCJnZXQ6ZHJpbmtkZXRhaWxzIiwiZ2V0OmRyaW5rcyIsInBhdGNoOmRyaW5rcyIsInBvc3Q6ZHJpbmtzIl19.y8sC8p0ffT68b4i3g8iTCjcF9LsZNtwv2hfjjxPh4F7tHZqpNqZOfWuOBzo-623XHxK8HKSTielujijG79F8E_rYiMfFprN-yhlgXbZokkRSGWb64SEUPVUEdhd5VqVx_Bm5Kx1AkHex924K4pZkpvHhjoRWHKVaOFt6iHq6Wrl9bX7x27PP0ytDh6zAC5XPvrzuVsUbX3JIho8M2VQeqhEEGy7lZ-_rI9BcCQ6Pfb0ZZJFRBRyMjMhsFt2Ihj6cl6EtdLYaSyBZYdeIKrH0XKchqg8Bu2iYh7aXHqLxa192DqKyY04LneCn_qM3f6mFVhf_68CpGFC95aRSEgfyFw
Barista: eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCIsImtpZCI6IlJ2d001NV9jdzEwdTZ5Sk9aNlMwNCJ9.eyJpc3MiOiJodHRwczovL2NvZmZlZS1kZXYudXMuYXV0aDAuY29tLyIsInN1YiI6ImF1dGgwfDYyYmMzMDRkNTY5NTJjZDNlZGFiMzdkNSIsImF1ZCI6ImNvZmZlZXNob3AiLCJpYXQiOjE2NTY1MDIzMDksImV4cCI6MTY1NjUwOTUwOSwiYXpwIjoiVzZ1V2hIYkt5RDdmSzE5YzRYS0JHTHROTDdDaHFha2kiLCJzY29wZSI6IiIsInBlcm1pc3Npb25zIjpbImdldDpkcmlua2RldGFpbHMiLCJnZXQ6ZHJpbmtzIl19.p8AD2oUdM3N84OzKsYj-ohG3MPnCYDPXRCcZy8NPCwbb7VFiIDQzjBgAhAVw4_Vr1yyEq6Lau_SGu9RlyNy1lIjHRJ30n4si1iVdmYpBAcQEh6Qui_oCIZom3u6bOj_Jwr15Jc1zzT2nk5nP3ievuI2cGnTASHdK4LaB-E-4__vH4Hg5MCfK0z7A35WW_ZP76mybH-ARaCYmWSk8TYTpTp0LdnABWluqyWXjRrYqb0Lm6ODAAaRkq78YF583MAg8FtMI9aBlVIAUP-kSpCu_Xy5AYmzuSw8lmwXPPANR_UkfNPsR2DEAnAzJJc9i0hHln0KpBU8PfI78IWYgAVm_Tw